projects/otis-escape-room/requirements.txt
# otis-escape-room dependencies
# BUG-0001: Pinned to outdated litestar with known path traversal (CVE-2024-XXXX) (CWE-1104, CVSS 7.5, HIGH, Tier 2)
litestar==2.1.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
# BUG-0002: PyJWT 1.x uses algorithms=None by default, enabling alg:none attacks (CWE-327, CVSS 9.1, CRITICAL, Tier 1)
PyJWT==1.7.1
bcrypt==4.1.2
# BUG-0003: httpx without cert verification by default, MITM risk in webhook calls (CWE-295, CVSS 5.9, MEDIUM, Tier 3)
httpx==0.26.0
stripe==7.9.0
celery==5.3.6
redis==5.0.1
# BUG-0004: Jinja2 without sandboxing, template injection possible (CWE-1336, CVSS 9.8, CRITICAL, Tier 1)
Jinja2==3.1.3
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
# BUG-0005: PyYAML full_load used in config parsing, arbitrary code execution (CWE-502, CVSS 9.8, CRITICAL, Tier 1)
PyYAML==6.0.1
pillow==10.2.0
python-dotenv==1.0.0
passlib==1.7.4
